handle tests on one command
