+++ manage.py
@@ -1,5 +1,6 @@
 
 import os
+import unittest
 from flask_script import Manager # class for handling a set of commands
 from flask_migrate import Migrate, MigrateCommand
 from app import db, create_app
